python/requirements.txt
websockets>=12.0
pyyaml>=6.0
psutil>=5.9
customtkinter>=5.2
pillow>=10.0
